added helper method to fix inserts error. will apply to possibly every future object
before, whenever I inputted "3" and tried to insert a new genre, it would cause an "exception unhandled" error. This happened because the database was still requiring GenreID to be manually provided rather than it being auto incremented. to fix this, I'd have to go back to my dbo.genre and set "is identity" to true, which sounds simple on paper, but it proved to be very difficult as I've already finalized the tables/schema of my database for a while now, and changing something in the table now would mean possibly having to erase all the data I've inserted and go through the create/drop table process all over again which i just didn't feel like doing. so instead, i opted for a helper method, which is an easy fix to this using just code and not having to modify the actual database.