Updated view all records to join artist and genre table in the query
I added a new method to show records with real names for foreign columns like artist and genre rather than just displaying the IDS. However I still left the old method that I was using to display records as there are some other parts of my app that still use it. However, I should be able to remove it once everything is updated to the new version, I should be able to phase that out eventually. But for now it stays.

Also, you may or may not have noticed that I now have List<string[]> instead of List<Record>, and this is because we're dealing with much more complicated data, like foreign columns/joined fields, and using string[lets me show joined table info without needing a new class.

I was also going to make the filtering/sorting options into sub options for the "view all" option, but I'm not sure about that yet.