User table alter/role based filtering + more
Switched from using IsAdmin to a Role column to better align with the criteria of role based filtering, so the system now checks the user's role instead of just a true/false if the person is admin or not.

I also added navigation guides/specification for the pagination, which is something I should've done a long time ago.

I also used SELECT SCOPE_IDENTITY(); so that after inserting a new user we'd immediately get the ID of that user (the one that was just added)