Added input validation for all numeric/int fields
 Changed all numeric inputs from view.GetIntInput to simply view.Getinput.

Used tryparse for user input conversion and also to prevent formatexception crashes during ID, year, or other numerical inputs during insertions.

Reason why I changed it from getintinput to simply getinput is because the user could just type something that isn't a number or leave it blank which leads to exception unhandled. Relying on GetInput in this case allows the user's input to be safely checked with tryparse first as a string before being converted into a proper integer to avoid crashes/exception unhandled.